+++ classes/taxonomies/wpdk-terms.php
@@ -28,8 +28,8 @@
  * @class           WPDKTerm
  * @author          =undo= <info@wpxtre.me>
  * @copyright       Copyright (C) 2012-2013 wpXtreme Inc. All Rights Reserved.
- * @date            2013-08-29
- * @version         1.0.1
+ * @date            2013-09-17
+ * @version         1.0.2
  *
  */
 class WPDKTerm {
@@ -50,6 +50,8 @@ class WPDKTerm {
    * Create an instance of WPDKTerm class
    *
    * @brief Construct
+   *
+   * @param object $term A WordPress term object
    *
    * @return WPDKTerm
    */